+++ src/views/examples/ProfilePage.js
@@ -25,11 +25,13 @@ function ProfilePage() {
     document.body.classList.add("profile-page");
     document.body.classList.add("sidebar-collapse");
     document.documentElement.classList.remove("nav-open");
+    window.scrollTo(0, 0);
+    document.body.scrollTop = 0;
     return function cleanup() {
       document.body.classList.remove("profile-page");
       document.body.classList.remove("sidebar-collapse");
     };
-  });
+  },[]);
   return (
     <>
       <ExamplesNavbar />